refactor: GameResult 패키지 이동
